# Conflicts:
#	FluentTerminal.App.Services/Constants.cs
#	FluentTerminal.App.Services/Implementation/TrayProcessCommunicationService.cs
#	FluentTerminal.Models/Responses/CreateTerminalResponse.cs
#	FluentTerminal.SystemTray/Services/AppCommunicationService.cs
#	FluentTerminal.SystemTray/Services/TerminalsManager.cs

+++ FluentTerminal.SystemTray/Services/WinPty/WinPtySession.cs
@@ -112,7 +112,7 @@ namespace FluentTerminal.SystemTray.Services.WinPty
 
         public event EventHandler<int> ConnectionClosed;
 
-        public int Id { get; private set; }
+        public byte Id { get; private set; }
 
         public string ShellExecutableName { get; private set; }
